Delete empty vendor directories after deleting package directories.
If "delete_vendor_directories" is set to true and all the packages
from a given vendor are deleted, the directory having the vendor name
inside "vendor" will still exist and be empty. This commit checks
if that's the case and then deletes the (now useless) empty
vendor directory.

Also, "is_link" was mistakenly using the relative package directory
prepended with "/", so it was being interpreted as absolute.
This commit fixes that too.